move `oracle` module to new `scarb-oracle-hint-processor` package
