--- extensions/scarb-execute/src/oracle/mod.rs
@@ -1,4 +1,0 @@
-mod connection;
-mod connections;
-pub mod hint_processor;
-mod jsonrpc;
