+++ utils/scarb-oracle-hint-processor/Cargo.toml
@@ -13,5 +13,14 @@ readme = "README.md"
 repository.workspace = true
 
 [dependencies]
+anyhow.workspace = true
+cairo-lang-casm.workspace = true
+cairo-lang-runner.workspace = true
+cairo-vm.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+shell-words.workspace = true
+starknet-core.workspace = true
+tracing.workspace = true
 
 [dev-dependencies]
